extracted scope's model, context, provider and use-scope query to they're own components
