+++ scopes/scope/models/scope-model/index.ts
