+++ scopes/scope/models/scope-model/scope.model.ts
