+++ scopes/scope/scope/scope.ui.runtime.tsx
@@ -12,10 +12,10 @@ import { RouteProps } from 'react-router-dom';
 import { MenuWidget, MenuWidgetSlot } from '@teambit/ui-foundation.ui.menu';
 import { MenuLinkItem } from '@teambit/design.ui.surfaces.menu.link-item';
 import CommandBarAspect, { CommandBarUI, ComponentSearcher, CommandHandler } from '@teambit/command-bar';
+import { ScopeModel } from '@teambit/scope.models.scope-model';
 import { ScopeMenu, ScopeUseBox } from './ui/menu';
 import { ScopeAspect } from './scope.aspect';
 import { Scope } from './ui/scope';
-import { ScopeModel } from './ui/scope-model';
 import { ComponentsDrawer } from './ui/components-drawer';
 
 export type ScopeBadge = ComponentType;
