--- scopes/scope/scope/ui/scope-context.ts
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-import { ScopeModel } from './scope-model';
-
-export const ScopeContext: React.Context<ScopeModel> = createContext<ScopeModel>(ScopeModel.empty());
