+++ scopes/scope/ui/hooks/scope-context/index.ts
@@ -1,0 +1,2 @@
+export { ScopeContext, useScope } from './scope-context';
+export { ScopeProvider } from './scope-provider';
