+++ scopes/scope/ui/hooks/use-scope/index.ts
@@ -1,0 +1,1 @@
+export { useScopeQuery } from './use-scope';
